Move interfaces into seperate path
